use-cases/01-medical-transcription-realtime/python/requirements.txt
# Direct dependencies only

python-dotenv>=1.2.1
speechmatics-rt>=0.5.0
typing_extensions>=4.0.0
